requirements.txt
crewai==0.51.1
crewai_tools==0.12.1
langchain_groq==0.1.5
langchain_core==0.2.41
langchain==0.2.16
langchain_community==0.2.17
langchain_experimental==0.0.65
python-dotenv